Improved dumper script and a Kaitai Struct spec.
Improved the spec for PocketBook swupdate.bin:
Added an additional enum value for uboot.
Added some metadata.
Renamed the file to match the id.

Improved dumper script:
Replaced reading file into memory with with memory mapping.
Added a mode to not to name dumps using hashes, but name with simple numbers.
Files are placed relative to the unpacked image.
Added shebang and made executable.
Reordered the imports with isort.